Works (read me)
Works, but not as expected. max_allocated_memory and smi show different
answers. So I can't trust the numbers. But SMI seems better because at
least it is closer to when the memory fails.

I just expanded the model until it wouldn't fit in memory. It was 300MB
over. I found that a 2 split didn't fit (gave same over allocation error
with exact same memory error). Then 4, 8, 10, 16. What is interesting is
that memory usage is 8 < 10 < 4 < 16. 16 Barely fits in memory.